+++ pkg/extension/opampextension/opamp_agent_test.go
@@ -1,0 +1,163 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package opampextension
+
+import (
+	"context"
+	"os"
+	"path/filepath"
+	"testing"
+
+	"github.com/oklog/ulid/v2"
+	"github.com/open-telemetry/opamp-go/protobufs"
+	"github.com/stretchr/testify/assert"
+	"go.opentelemetry.io/collector/component/componenttest"
+	"go.opentelemetry.io/collector/extension/extensiontest"
+)
+
+func TestNewOpampAgent(t *testing.T) {
+	cfg := createDefaultConfig()
+	set := extensiontest.NewNopCreateSettings()
+	o, err := newOpampAgent(cfg.(*Config), set.Logger)
+	assert.NoError(t, err)
+	assert.NotEmpty(t, o.instanceId.String())
+	assert.Empty(t, o.effectiveConfig)
+	assert.Nil(t, o.agentDescription)
+}
+
+func TestCreateAgentDescription(t *testing.T) {
+	cfg := createDefaultConfig()
+	set := extensiontest.NewNopCreateSettings()
+	o, err := newOpampAgent(cfg.(*Config), set.Logger)
+	assert.NoError(t, err)
+
+	assert.Nil(t, o.agentDescription)
+	assert.NoError(t, o.createAgentDescription())
+	assert.NotNil(t, o.agentDescription)
+}
+
+func TestLoadEffectiveConfig(t *testing.T) {
+	cfg := createDefaultConfig()
+	set := extensiontest.NewNopCreateSettings()
+	o, err := newOpampAgent(cfg.(*Config), set.Logger)
+	assert.NoError(t, err)
+
+	assert.Empty(t, o.effectiveConfig)
+
+	path := filepath.Join("testdata", "opamp-remote-config.yaml")
+	assert.NoError(t, o.loadEffectiveConfig(path))
+	assert.NotEmpty(t, o.effectiveConfig)
+}
+
+func TestSaveEffectiveConfig(t *testing.T) {
+	cfg := createDefaultConfig()
+	set := extensiontest.NewNopCreateSettings()
+	o, err := newOpampAgent(cfg.(*Config), set.Logger)
+	assert.NoError(t, err)
+
+	f, err := os.CreateTemp("", "opamp-remote-config.yaml")
+	assert.NoError(t, err)
+	defer os.Remove(f.Name())
+
+	assert.NoError(t, o.saveEffectiveConfig(f.Name()))
+}
+
+func TestUpdateAgentIdentity(t *testing.T) {
+	cfg := createDefaultConfig()
+	set := extensiontest.NewNopCreateSettings()
+	o, err := newOpampAgent(cfg.(*Config), set.Logger)
+	assert.NoError(t, err)
+
+	olduid := o.instanceId
+	assert.NotEmpty(t, olduid.String())
+
+	uid := ulid.Make()
+	assert.NotEqual(t, uid, olduid)
+
+	o.updateAgentIdentity(uid)
+	assert.Equal(t, o.instanceId, uid)
+}
+
+func TestComposeEffectiveConfig(t *testing.T) {
+	cfg := createDefaultConfig()
+	set := extensiontest.NewNopCreateSettings()
+	o, err := newOpampAgent(cfg.(*Config), set.Logger)
+	assert.NoError(t, err)
+
+	ec := o.composeEffectiveConfig()
+	assert.NotNil(t, ec)
+}
+
+func TestApplyRemoteConfig(t *testing.T) {
+	cfg := createDefaultConfig()
+	set := extensiontest.NewNopCreateSettings()
+	o, err := newOpampAgent(cfg.(*Config), set.Logger)
+	assert.NoError(t, err)
+
+	assert.Empty(t, o.effectiveConfig)
+
+	path := filepath.Join("testdata", "opamp-remote-config.yaml")
+	rb, err := os.ReadFile(path)
+	assert.NoError(t, err)
+
+	rc := &protobufs.AgentRemoteConfig{
+		Config: &protobufs.AgentConfigMap{
+			ConfigMap: map[string]*protobufs.AgentConfigFile{
+				"default": {
+					Body: rb,
+				},
+			},
+		},
+		ConfigHash: []byte("b2b1e3e7f45d564db1c0b621bbf67008"),
+	}
+
+	changed, err := o.applyRemoteConfig(rc)
+	assert.True(t, changed)
+	assert.NoError(t, err)
+	assert.NotEmpty(t, o.effectiveConfig)
+}
+
+func TestShutdown(t *testing.T) {
+	cfg := createDefaultConfig().(*Config)
+	cfg.HTTPClientSettings.Auth = nil
+	set := extensiontest.NewNopCreateSettings()
+	o, err := newOpampAgent(cfg, set.Logger)
+	assert.NoError(t, err)
+
+	// Shutdown with no OpAMP client
+	assert.NoError(t, o.Shutdown(context.Background()))
+}
+
+func TestStart(t *testing.T) {
+	cfg := createDefaultConfig().(*Config)
+	cfg.HTTPClientSettings.Auth = nil
+	set := extensiontest.NewNopCreateSettings()
+	o, err := newOpampAgent(cfg, set.Logger)
+	assert.NoError(t, err)
+
+	assert.NoError(t, o.Start(context.Background(), componenttest.NewNopHost()))
+}
+
+func TestReload(t *testing.T) {
+	cfg := createDefaultConfig().(*Config)
+	cfg.HTTPClientSettings.Auth = nil
+	set := extensiontest.NewNopCreateSettings()
+	o, err := newOpampAgent(cfg, set.Logger)
+	assert.NoError(t, err)
+
+	ctx := context.Background()
+	assert.NoError(t, o.Start(ctx, componenttest.NewNopHost()))
+	assert.NoError(t, o.Reload(ctx))
+}